jxl-bitstream: Reorganize modules
